:space_invader: moved to utilfunctions / added test to package
