+++ .travis.yml
@@ -2,5 +2,4 @@ language: node_js
 node_js:
   - "lts/carbon"
 install:
-  - "npm install -g mocha"
   - "npm install"
